refactor(interactive-section): extract section-numbering module
Phase: Tier A
Step: A2
Pattern: none (pure helper extraction)
Risk: low

Why: shouldNumberSectionChild and wrapSectionChildrenForNumbering are
pure helpers with no module-level state and no behaviour dependent on
section instance. They sit in interactive-section.tsx purely for
historical reasons. Extracting them shrinks the section component
file, gives the helpers their own colocated home next to the existing
section-numbering.test.tsx, and removes the docs-retrieval +
interactive-conditional imports from interactive-section.tsx (those
component-type identities are only used for the numbering predicate).

interactive-section.tsx re-exports both functions so the existing
section-numbering.test.tsx (which imports from ./interactive-section)
continues to work without modification. No consumer migration in this
commit per Principle 6.

Cycle-safety preserved: the comment about call-time component-identity
lookups moves into section-numbering.tsx unchanged, because the same
docs-retrieval barrel ↔ content-renderer ↔ interactive-tutorial cycle
that motivated the original behaviour still exists at the new file
location.

Checkpoint:
  - npm run typecheck: pass
  - npx eslint: pass
  - npx prettier --check: pass
  - npx jest src/components/interactive-tutorial: 16 suites pass,
    246 tests pass + 5 todos + 12 pre-existing skips
  - section-numbering.test.tsx continues to pass via re-exports.
  - interactive-section.tsx line delta: -54 / +5 (~49 lines reduced).

Notes:
  - The file is intentionally .tsx (not .ts) because
    wrapSectionChildrenForNumbering returns JSX (an `<li>` element).
  - No test file is added — the existing section-numbering.test.tsx
    already covers both functions and continues to import via the
    re-export, so the move is implicit Pre-test → Extract →
    Post-test = same-test in this case (Principle 4).
